Extract OpenMetricsTextUtils
